moved utils method in utils file.
